+++ lib/js_of_ocaml/dom_html.ml
@@ -3286,6 +3286,8 @@ class type document = object
 
   method images : imageElement collection t readonly_prop
 
+  method fonts : FontFace.fontFaceSet t readonly_prop
+
   method applets : element collection t readonly_prop
 
   method links : element collection t readonly_prop